USAGOV-1997-mobile-menu: Fixing lint errors
